requirements.txt
Python libraries required for the project

discord.py>=2.3.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0